+++ lockfile/src/csharp.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use std::ffi::OsStr;
 use std::path::Path;
 
@@ -7,9 +8,75 @@ use serde_xml_rs::Deserializer;
 
 use crate::{Package, PackageVersion, Parse};
 
-pub struct CSProj;
-
 const UTF8_BOM: &str = "\u{feff}";
+
+pub struct PackagesLock;
+
+impl Parse for PackagesLock {
+    /// Parses `packages.lock.json` files into a vec of packages
+    fn parse(&self, data: &str) -> anyhow::Result<Vec<Package>> {
+        // Deserialize lockfile as JSON.
+        let parsed: PackagesLockJson = serde_json::from_str(data)?;
+
+        // Map all dependencies to their correct package types.
+        let packages = parsed
+            .dependencies
+            .into_iter()
+            .flat_map(|(_, deps)| deps.into_iter())
+            .map(|(name, dependency)| {
+                let version = match dependency.dependency_type {
+                    DependencyType::Project => PackageVersion::Path(None),
+                    _ => PackageVersion::FirstParty(dependency.resolved.unwrap()),
+                };
+
+                Package { version, name, package_type: PackageType::Nuget }
+            })
+            .collect();
+
+        Ok(packages)
+    }
+
+    fn is_path_lockfile(&self, path: &Path) -> bool {
+        path.file_name() == Some(OsStr::new("packages.lock.json"))
+    }
+
+    fn is_path_manifest(&self, _path: &Path) -> bool {
+        false
+    }
+}
+
+/// `packages.lock.json` structure.
+#[derive(Deserialize, Debug)]
+struct PackagesLockJson {
+    #[serde(rename = "version")]
+    _version: usize,
+    dependencies: HashMap<String, FrameworkDependencies>,
+}
+
+/// `packages.lock.json` .NET framework structure.
+type FrameworkDependencies = HashMap<String, Dependency>;
+
+/// `packages.lock.json` dependency structure.
+#[derive(Deserialize, Debug)]
+struct Dependency {
+    #[serde(rename = "type")]
+    dependency_type: DependencyType,
+    resolved: Option<String>,
+}
+
+/// `packages.lock.json` dependency types.
+#[derive(Deserialize, Debug)]
+enum DependencyType {
+    Direct,
+    CentralTransitive,
+    Transitive,
+    Project,
+    Build,
+    Platform,
+    Tool,
+}
+
+pub struct CSProj;
 
 #[derive(Debug, Deserialize, PartialEq, Eq)]
 pub struct PackageReference {
@@ -79,6 +146,55 @@ impl Parse for CSProj {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn packages_lock() {
+        let pkgs =
+            PackagesLock.parse(include_str!("../../tests/fixtures/packages.lock.json")).unwrap();
+        assert_eq!(pkgs.len(), 7);
+
+        let expected_pkgs = [
+            Package {
+                name: "Microsoft.Windows.SDK.Contracts".into(),
+                version: PackageVersion::FirstParty("10.0.22621.755".into()),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "SSH.NET".into(),
+                version: PackageVersion::FirstParty("2020.0.2".into()),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "example.helpers".into(),
+                version: PackageVersion::Path(None),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "Microsoft.SourceLink.GitHub".into(),
+                version: PackageVersion::FirstParty("1.1.1".into()),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "Microsoft.Build.Tasks.Git".into(),
+                version: PackageVersion::FirstParty("1.1.1".into()),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "System.Buffers".into(),
+                version: PackageVersion::FirstParty("4.5.1".into()),
+                package_type: PackageType::Nuget,
+            },
+            Package {
+                name: "Microsoft.CodeAnalysis.FxCopAnalyzers".into(),
+                version: PackageVersion::FirstParty("3.3.0".into()),
+                package_type: PackageType::Nuget,
+            },
+        ];
+
+        for expected_pkg in expected_pkgs {
+            assert!(pkgs.contains(&expected_pkg), "missing package {expected_pkg:?}");
+        }
+    }
 
     #[test]
     fn lock_parse_csproj() {